+++ oca/agents/acpg.py
@@ -16,7 +16,7 @@ from torch.types import Number, Tensor
 
 from ..envs.fourrooms import FourRoomsEnv
 from ..envs.utils import OneHotWrapper
-from ..utils.config import ConfigRunProto
+from ..utils.config import ConfigEvalProto, ConfigRunProto
 from ..utils.constants import models_dir
 from ..utils.logger import RegularLogger
 
@@ -107,7 +107,7 @@ class ActorCriticAgent(nn.Module):
         return total_reward
 
 
-def run_acpg(args: ConfigRunProto, env: gym.Env, **kwargs):
+def run(args: ConfigRunProto, env: gym.Env, **kwargs):
     state_dim: int = env.observation_space.shape[0]  # if one-hot state
     action_dim: int = env.action_space.n
     env = OneHotWrapper(env)
@@ -152,3 +152,7 @@ def run_acpg(args: ConfigRunProto, env: gym.Env, **kwargs):
         env.render()
 
         logger.log_episode(ep, reward)
+
+
+def evaluate(args: ConfigEvalProto, env: gym.Env, **kwargs):
+    raise NotImplementedError