+++ oca/agents/sarsa.py
@@ -16,7 +16,7 @@ from torch import Tensor
 
 from ..envs.fourrooms import FourRoomsEnv
 from ..envs.utils import OneHotWrapper
-from ..utils.config import ConfigRunProto
+from ..utils.config import ConfigEvalProto, ConfigRunProto
 from ..utils.constants import models_dir
 from ..utils.logger import RegularLogger
 
@@ -103,7 +103,7 @@ class SARSAAgent(nn.Module):
         return total_reward
 
 
-def run_sarsa(args: ConfigRunProto, env: gym.Env, **kwargs):
+def run(args: ConfigRunProto, env: gym.Env, **kwargs):
     state_dim: int = env.observation_space.shape[0]  # if one-hot state
     action_dim: int = env.action_space.n
     env = OneHotWrapper(env)
@@ -147,3 +147,7 @@ def run_sarsa(args: ConfigRunProto, env: gym.Env, **kwargs):
         env.render()
 
         logger.log_episode(ep, reward)
+
+
+def evaluate(args: ConfigEvalProto, env: gym.Env, **kwargs):
+    raise NotImplementedError